+++ ZGO2/include/model/go/goanalysis.h
@@ -17,7 +17,7 @@ public:
 
 protected:
     bool isCommonSignalAppeared(GOOperator *op, int index, QVector<GOOperator *> commonOperator, QVector<int> commonIndex);
-    double calcTempAccumulative(GOOperator *op, int index, QVector<GOOperator *> &commonOperator, QVector<int> &commonIndex, QVector<double> &accValues, int accIndex);
+    long double calcTempAccumulative(GOOperator *op, int index, QVector<GOOperator *> &commonOperator, QVector<int> &commonIndex, QVector<long double> &accValues, int accIndex);
     virtual void calcAccumulativeProbabilityType1(GOOperator *op);
     virtual void calcAccumulativeProbabilityType2(GOOperator *op);
     virtual void calcAccumulativeProbabilityType3(GOOperator *op);
@@ -36,23 +36,23 @@ protected:
     virtual void calcAccumulativeProbabilityType16(GOOperator *op);
     virtual void calcAccumulativeProbabilityType17(GOOperator *op);
 
-    virtual double calcTempAccumulativeType1(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType2(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType3(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType4(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType5(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType6(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType7(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType8(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType9(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType10(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType11(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType12(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType13(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType14(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType15(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType16(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
-    virtual double calcTempAccumulativeType17(GOOperator *op, QVector<double> inputValues, QVector<double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType1(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType2(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType3(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType4(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType5(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType6(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType7(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType8(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType9(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType10(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType11(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType12(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType13(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType14(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType15(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType16(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
+    virtual long double calcTempAccumulativeType17(GOOperator *op, QVector<long double> inputValues, QVector<long double> subInputValues, int accIndex);
 };
 
 #endif // GOANALYSIS_H
